Connected the frontend to the backend for updating a unit review.
profile.component passes write-review-unit.component the unit id and the review object itself.
write-review-unit.component processes the review and sends the review object to the backend via API service.
write-review-unit.component has a editReview function which is only called if the variable editMode is true.
It sends the review object to the backend using API editReviewPUT.
API service then sends the review object to the backend.